# contracts/openapi/gateway-node.openapi.yaml
openapi: 3.1.0
info:
  title: Ishtirak gateway (public BFF)
  version: 0.1.0
  description: >
    The public API the web client consumes (ADR-003). The gateway authenticates
    every request (Bearer JWT), derives operatorId + role from the token, and
    aggregates/proxies to the internal core-java and analytics-python services.
    All resources are implicitly scoped to the caller's operator — clients never
    pass operatorId.


    The real-time channel (`GET /api/ws`, WebSocket upgrade) is **not** modeled
    in OpenAPI; its protocol lives in ../asyncapi.yaml and ../../docs/API.md.


    **Design-ahead-of-implementation:** only `/health` and `/ready` are built
    today (`x-status: built`); everything else is `x-status: planned`.
  contact: { name: Ishtirak, url: https://ishtirak.dev }
servers:
  - url: http://localhost:8080
    description: Local / docker-compose (public)
security:
  - bearerAuth: []
...
paths:
  /api/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Exchange email/password credentials
      description: >
        Rate-limited (brute-force / credential-stuffing protection). The
        `password` field MUST never be logged (the gateway uses pino-http
        request logging — redact this body). Users are global by email and may
        have multiple operator memberships; multi-membership logins return a
        short-lived selection token instead of context-bound access tokens.
      x-status: planned
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password:
                  type: string
                  format: password
                  description: Never logged or echoed back.
      responses:
        '200':
          description: Authenticated or context selection required.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [contextSelectionRequired, memberships]
                properties:
                  contextSelectionRequired: { type: boolean }
                  selectionToken: { type: string }
                  accessToken: { type: string }
                  refreshToken: { type: string }
                  memberships:
                    type: array
                    items:
                      type: object
                      additionalProperties: false
                      required: [membershipId, operatorId, operatorName, role]
                      properties:
                        membershipId: { type: string, format: uuid }
                        operatorId: { type: string, format: uuid }
                        operatorName: { type: string }
                        role: { type: string, enum: [OPERATOR_ADMIN, OPERATOR_STAFF, SUBSCRIBER] }
                        subscriberId: { type: string, format: uuid }
        '400': { $ref: 'components.yaml#/components/responses/BadRequest' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '429': { $ref: 'components.yaml#/components/responses/TooManyRequests' }
  /api/auth/select-context:
    post:
      tags: [auth]
      operationId: selectAuthContext
      summary: Select an operator membership
      x-status: planned
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [selectionToken, membershipId]
              properties:
                selectionToken: { type: string }
                membershipId: { type: string, format: uuid }
      responses:
        '200':
          description: Context-bound tokens.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [accessToken, refreshToken]
                properties:
                  accessToken: { type: string }
                  refreshToken: { type: string }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
  /api/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Refresh an access token
      description: >
        Refresh tokens are single-use and rotated on each refresh; detecting reuse
        of a consumed token revokes the token family (theft mitigation).
        Rate-limited.
      x-status: planned
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New access token (and rotated refresh token).
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [accessToken]
                properties:
                  accessToken: { type: string }
                  refreshToken: { type: string }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '429': { $ref: 'components.yaml#/components/responses/TooManyRequests' }
...
  /api/billing-runs:
    post:
      tags: [billing]
      operationId: apiRunBilling
      summary: Trigger the monthly billing run
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN]
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [periodStart, periodEnd]
              properties:
                periodStart: { type: string, format: date }
                periodEnd: { type: string, format: date }
      responses:
        '202':
          description: Billing run accepted.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [issued, periodStart, periodEnd]
                properties:
                  issued: { type: integer, minimum: 0 }
                  periodStart: { type: string, format: date }
                  periodEnd: { type: string, format: date }
        '400': { $ref: 'components.yaml#/components/responses/BadRequest' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
        '409': { $ref: 'components.yaml#/components/responses/Conflict' }
  /api/invoices:
    get:
      tags: [billing]
      operationId: apiListInvoices
      summary: List invoices
      x-status: planned
      parameters:
        - { name: status, in: query, required: false, schema: { type: string, enum: [ISSUED, PARTIAL, PAID, VOID] } }
        - { name: periodStart, in: query, required: false, schema: { type: string, format: date } }
        - { name: periodEnd, in: query, required: false, schema: { type: string, format: date } }
        - $ref: 'components.yaml#/components/parameters/pageParam'
        - $ref: 'components.yaml#/components/parameters/limitParam'
      responses:
        '200':
          description: Page of invoices.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: 'components.yaml#/components/schemas/Invoice' }
                  meta: { $ref: 'components.yaml#/components/schemas/PageMeta' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
  /api/invoices/{id}:
    parameters:
      - $ref: 'components.yaml#/components/parameters/idParam'
    get:
      tags: [billing]
      operationId: apiGetInvoice
      summary: Get an invoice
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN, OPERATOR_STAFF]
      responses:
        '200':
          description: Invoice.
          content:
            application/json:
              schema: { $ref: 'components.yaml#/components/schemas/Invoice' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
        '404': { $ref: 'components.yaml#/components/responses/NotFound' }
  /api/invoices/{id}/payments:
    parameters:
      - $ref: 'components.yaml#/components/parameters/idParam'
    get:
      tags: [payments]
      operationId: apiListInvoicePayments
      summary: Payments recorded against an invoice
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN, OPERATOR_STAFF]
      responses:
        '200':
          description: Payments for the invoice.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: 'components.yaml#/components/schemas/Payment' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
        '404': { $ref: 'components.yaml#/components/responses/NotFound' }
  /api/me/invoices:
    get:
      tags: [billing]
      operationId: apiListMyInvoices
      summary: The caller's own invoices (subscriber portal)
      x-status: planned
      x-required-roles: [SUBSCRIBER]
      parameters:
        - $ref: 'components.yaml#/components/parameters/pageParam'
        - $ref: 'components.yaml#/components/parameters/limitParam'
      responses:
        '200':
          description: Page of the subscriber's invoices.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: 'components.yaml#/components/schemas/Invoice' }
                  meta: { $ref: 'components.yaml#/components/schemas/PageMeta' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
...
  /api/me/invoices/{id}:
    parameters:
      - $ref: 'components.yaml#/components/parameters/idParam'
    get:
      tags: [billing]
      operationId: apiGetMyInvoice
      summary: One invoice owned by the subscriber caller
      x-status: planned
      x-required-roles: [SUBSCRIBER]
      responses:
        '200':
          description: Invoice.
          content:
            application/json:
              schema: { $ref: 'components.yaml#/components/schemas/Invoice' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
        '404': { $ref: 'components.yaml#/components/responses/NotFound' }
  /api/me/invoices/{id}/payments:
    parameters:
      - $ref: 'components.yaml#/components/parameters/idParam'
    get:
      tags: [payments]
      operationId: apiListMyInvoicePayments
      summary: Payments for one invoice owned by the subscriber caller
      x-status: planned
      x-required-roles: [SUBSCRIBER]
      responses:
        '200':
          description: Payments for the invoice.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: 'components.yaml#/components/schemas/Payment' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
        '404': { $ref: 'components.yaml#/components/responses/NotFound' }
  /api/payments:
    post:
      tags: [payments]
      operationId: apiRecordPayment
      summary: Record a payment against an invoice
      description: Proxies to core-java, which emits `payment.received`.
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN, OPERATOR_STAFF]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [invoiceId, currency, tenderedAmount, method]
              properties:
                invoiceId: { type: string, format: uuid }
                currency: { $ref: 'components.yaml#/components/schemas/Currency' }
                tenderedAmount: { type: number, minimum: 0 }
                method: { type: string, enum: [CASH, WHISH] }
      responses:
        '201':
          description: Recorded payment.
          content:
            application/json:
              schema: { $ref: 'components.yaml#/components/schemas/Payment' }
        '400': { $ref: 'components.yaml#/components/responses/BadRequest' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
        '404': { $ref: 'components.yaml#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      $ref: 'components.yaml#/components/securitySchemes/bearerAuth'